document-printing-service/requirements.txt
Flask==2.3.3
pymongo==4.11.1
certifi==2026.1.4
mongomock==4.3.0
python-dotenv==1.0.0
Werkzeug==2.3.7
dnspython==2.4.2
Pillow==10.4.0
PyMuPDF==1.24.10
python-docx==1.1.2
razorpay==1.4.2
APScheduler==3.10.4
gunicorn==23.0.0
